\name{checkLabel}
\alias{checkLabel}
\title{Checks a tree for a valid label, optionally removing bad characters on the way. This is currently a no-op,
to avoid interfering with the NHX annotation support.}
\usage{
  checkLabel(x, ...)
}
\arguments{
  \item{x,}{string input}
}
\value{
  string
}
\description{
  Checks a tree for a valid label, optionally removing bad
  characters on the way. This is currently a no-op, to
  avoid interfering with the NHX annotation support.
}

