\docType{data}
\name{GeomJoinedSegment}
\alias{GeomJoinedSegment}
\title{Create a geom_joinedsegment which joins together segments with 'round' corners. Looks MUCH nicer.
See https://groups.google.com/forum/?fromgroups#!topic/ggplot2/movv0f_MSuY}
\format{proto object 
 $ draw   :function (., data, scales, coordinates, arrow = NULL, ...)   
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 639 10 657 1 10 1 639 657 
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x105bb6f78>  
 $ objname: chr "geom_joinedsegment" 
 parent: proto object 
 .. parent: proto object 
 .. .. parent: proto object}
\usage{
  GeomJoinedSegment
}
\description{
  Create a geom_joinedsegment which joins together segments
  with 'round' corners. Looks MUCH nicer. See
  https://groups.google.com/forum/?fromgroups#!topic/ggplot2/movv0f_MSuY
}
\keyword{datasets}

