\docType{package}
\name{ggphylo}
\alias{ggphylo}
\alias{ggphylo-package}
\alias{package-ggphylo}
\title{Plots a tree or a list of trees.}
\usage{
  ggphylo(x, layout = "default", do.plot = T,
    plot.internals = F, x.lab = "Branch Length",
    x.lim = NA, x.expand = c(0.05, 0.3), y.lim = NA, ...)
}
\description{
  Plots a tree or a list of trees.

  \code{ggphylo} provides convenient functions and tools
  for visualizing phylogenetic data in R. Many methods are
  provided to simplify the process of working with
  \code{phylo} objects (e.g., \code{\link{tree.scale.to}}
  and \code{\link{tree.translate}} , while two plotting
  methods allow for a wide range of expressive plots:
  \code{\link{tree.plot}} and \code{\link{aln.plot}}
}

