\name{is.tree}
\alias{is.tree}
\title{Returns TRUE if the given object appears to be a valid tree object for the purposes of this package.}
\usage{
  is.tree(x)
}
\arguments{
  \item{x,}{input phylo object}
}
\value{
  boolean, TRUE if it is a tree, FALSE if not
}
\description{
  Returns TRUE if the given object appears to be a valid
  tree object for the purposes of this package.
}

