\name{label.phylo}
\alias{label.phylo}
\title{Returns the label for the given node. Alias for \code{\link{tree.label.for.node}}. Given
a nice short nickname due to its usefulness.}
\usage{
  \method{label}{phylo} (phylo, node)
}
\arguments{
  \item{phylo,}{input phylo object}

  \item{node,}{the node index for the desired label}
}
\description{
  Returns the label for the given node. Alias for
  \code{\link{tree.label.for.node}}. Given a nice short
  nickname due to its usefulness.
}

