\name{labels.phylo}
\alias{labels.phylo}
\title{Returns labels corresponding to all or some subset of the nodes in a tree.}
\usage{
  \method{labels}{phylo} (phylo, nodes = NA)
}
\arguments{
  \item{phylo}{input phylo object}

  \item{nodes}{[optional] vector of node indices for which
  to return labels}
}
\description{
  Returns labels corresponding to all or some subset of the
  nodes in a tree.
}

