\name{leaves}
\alias{leaves}
\title{Returns a vector of node indices for all the leaves in the tree.}
\usage{
  leaves(phylo)
}
\arguments{
  \item{phylo}{input phylo object}
}
\value{
  integer vector, with the node indices for all leaves
}
\description{
  Returns a vector of node indices for all the leaves in
  the tree.
}

