\name{nodes}
\alias{nodes}
\title{Returns a vector of node indices for all nodes (leaves plus internal nodes) in the tree.}
\usage{
  nodes(phylo)
}
\arguments{
  \item{phylo}{input phylo object}
}
\value{
  integer vector with the node indices for all nodes
}
\description{
  Returns a vector of node indices for all nodes (leaves
  plus internal nodes) in the tree.
}

