\name{tags.phylo}
\alias{tags.phylo}
\title{Retrieves a list of all tags for the given node. Convenience wrapper around \code{\link{tree.get.tags}}.}
\usage{
  \method{tags}{phylo} (phylo, node)
}
\description{
  Retrieves a list of all tags for the given node.
  Convenience wrapper around \code{\link{tree.get.tags}}.
}

