\name{tags.to.nhx.string}
\alias{tags.to.nhx.string}
\title{Converts a list of tags to an NHX annotation string.}
\usage{
  tags.to.nhx.string(tags)
}
\arguments{
  \item{tags}{list of tags}
}
\value{
  character containing [&&NHX:key=val] style annotations
  for all tags
}
\description{
  Converts a list of tags to an NHX annotation string.
}

