\name{theme_phylo_blank}
\alias{theme_phylo_blank}
\title{A mostly-blank theme for simple plots.
Ruthlessly taken from ggdendro, courtesy of Andrie de Vries
http://cran.r-project.org/web/packages/ggdendro/index.html}
\usage{
  theme_phylo_blank()
}
\description{
  A mostly-blank theme for simple plots. Ruthlessly taken
  from ggdendro, courtesy of Andrie de Vries
  http://cran.r-project.org/web/packages/ggdendro/index.html
}

