\name{tree.extract.clade}
\alias{tree.extract.clade}
\title{Returns the sub-tree beneath the given node.}
\usage{
  tree.extract.clade(phylo, node)
}
\arguments{
  \item{phylo,}{input phylo object}

  \item{node,}{the node from which to extract the clade /
  sub-tree}
}
\value{
  phylo, a new phylo object corresponding to the sub-tree
  beneath the given node.
}
\description{
  Returns the sub-tree beneath the given node.
}

