\name{tree.get.label}
\alias{tree.get.label}
\title{Returns the label for a given node in the tree. Alias for \code{\link{tree.label.for.node}}.}
\usage{
  tree.get.label(tree, node)
}
\arguments{
  \item{phylo}{input phylo object}

  \item{node}{integer, the node's index}
}
\value{
  the label for the given node
}
\description{
  Returns the label for a given node in the tree. Alias for
  \code{\link{tree.label.for.node}}.
}

