\name{tree.get.root}
\alias{tree.get.root}
\title{Returns the index of the root node in the tree, or NA if there is no root}
\usage{
  tree.get.root(phylo)
}
\arguments{
  \item{phylo}{input phylo object}
}
\value{
  integer vector corresponding to the index of the root
  node, or NA if there is no root.
}
\description{
  Returns the index of the root node in the tree, or NA if
  there is no root
}

