\name{tree.label.for.node}
\alias{tree.label.for.node}
\title{Returns the label for a given node in the tree.}
\usage{
  tree.label.for.node(tree, node)
}
\arguments{
  \item{phylo}{input phylo object}

  \item{node}{integer, the node's index}
}
\value{
  the label for the given node
}
\description{
  Returns the label for a given node in the tree.
}

