\name{tree.layout}
\alias{tree.layout}
\title{Returns a data frame defining segments to draw the phylogenetic tree.}
\usage{
  tree.layout(phylo, layout = "default",
    layout.ancestors = FALSE, align.seq.names = NA)
}
\description{
  Returns a data frame defining segments to draw the
  phylogenetic tree.
}

