\name{tree.leaves.beneath}
\alias{tree.leaves.beneath}
\title{Returns the number of leaves beneath a node in the tree (where a
leaf node is defined as having 1 leaf beneath)}
\usage{
  tree.leaves.beneath(phylo, node)
}
\arguments{
  \item{phylo}{input phylo object}

  \item{node}{integer, the index of the node to count
  leaves beneath}
}
\description{
  Returns the number of leaves beneath a node in the tree
  (where a leaf node is defined as having 1 leaf beneath)
}

