\name{tree.lengths.to.root}
\alias{tree.lengths.to.root}
\title{Returns a vector containing the root-to-tip branch length for each
leaf in the tree.}
\usage{
  tree.lengths.to.root(phylo)
}
\arguments{
  \item{phylo}{input phylo object}
}
\value{
  numeric vector, with one entry for each leaf node.
}
\description{
  Returns a vector containing the root-to-tip branch length
  for each leaf in the tree.
}

