\name{tree.max.length.to.root}
\alias{tree.max.length.to.root}
\title{Returns the maximum root-to-tip branch length in the tree. The
root-to-tip branch length is calculated for each node, and the
maximum value is returned.}
\usage{
  tree.max.length.to.root(phylo)
}
\arguments{
  \item{phylo}{input phylo object}
}
\description{
  Returns the maximum root-to-tip branch length in the
  tree. The root-to-tip branch length is calculated for
  each node, and the maximum value is returned.
}

