\name{tree.mean.path.length}
\alias{tree.mean.path.length}
\title{Returns the mean root-to-tip path length of the tree. This is
defined as the mean, across all leaves, of the total branch length
from the root to the leaf.}
\usage{
  tree.mean.path.length(phylo)
}
\arguments{
  \item{phylo}{input phylo object}
}
\value{
  numeric, the mean root-to-tip path length.
}
\description{
  Returns the mean root-to-tip path length of the tree.
  This is defined as the mean, across all leaves, of the
  total branch length from the root to the leaf.
}

