\name{tree.plot}
\alias{tree.plot}
\title{Plots a tree or a list of trees.}
\usage{
  tree.plot(x, plot.ancestors = T, axis.text.size = 5,
    keep.aspect.ratio = F, radial.plot = F, do.plot = T,
    show.y.gridlines = F, plot.legend = T, xlim = NA,
    xlim.expand = c(0.05, 0.3), xlab = "Branch Length",
    labels.in.axis = F, labels.in.tree = T,
    line.plot = TRUE, line.size = 1.5, line.size.by = NA,
    line.color.by = NA, node.plot = FALSE, node.size = 1,
    node.size.by = NA, node.color.by = NA,
    label.plot = TRUE, label.size = 4, label.size.by = NA,
    label.color.by = NA, y.lim = NA, ...)
}
\description{
  Plots a tree or a list of trees.
}

