\name{tree.remove.leaf}
\alias{tree.remove.leaf}
\title{Removes a single leaf from the tree and returns a copy.}
\usage{
  tree.remove.leaf(phylo, leaf)
}
\arguments{
  \item{phylo,}{input phylo object}

  \item{leaf,}{the node index of the desired leaf to
  remove.}
}
\value{
  phylo, a new phylo object with the altered tree
  structure.
}
\description{
  Removes a single leaf from the tree and returns a copy.
}

