\name{tree.scale.by}
\alias{tree.scale.by}
\title{Scales the tree by a given multiplicative factor.}
\usage{
  tree.scale.by(phylo, factor)
}
\arguments{
  \item{phylo}{input phylo object}

  \item{factor}{numeric, the factor to scale all branch
  lengths by}
}
\value{
  phylo, a new phylo object scaled by the desired amount.
}
\description{
  Scales the tree by a given multiplicative factor.
}

