\name{tree.scale.to}
\alias{tree.scale.to}
\title{Scales the tree to a given total branch length.}
\usage{
  tree.scale.to(phylo, total.length)
}
\arguments{
  \item{phylo}{input phylo object}

  \item{factor}{numeric, the desired total branch length of
  the tree}
}
\value{
  phylo, a new phylo object scaled by the required amount.
}
\description{
  Scales the tree to a given total branch length.
}

