\name{tree.set.label}
\alias{tree.set.label}
\title{Replaces the label of a given node and returns a new phylo object}
\usage{
  tree.set.label(tree, node, label)
}
\arguments{
  \item{phylo}{input phylo object}

  \item{node}{the index of the node whose label will be
  replaced}

  \item{label}{the new label for the given node}
}
\value{
  a new phylo object, with the node's label replaced
}
\description{
  Replaces the label of a given node and returns a new
  phylo object
}

