\name{tree.tags.to.labels}
\alias{tree.tags.to.labels}
\title{Converts all tags from a tree into NHX annotations and applies those annotations to each node's label.}
\usage{
  tree.tags.to.labels(phylo)
}
\arguments{
  \item{phylo}{input phylo object}
}
\value{
  a phylo object, where tags have been concatenated to each
  node's label
}
\description{
  Converts all tags from a tree into NHX annotations and
  applies those annotations to each node's label.
}

