\name{tree.to.string}
\alias{tree.to.string}
\title{Converts a phylo object into a Newick-format string representation. Taken mostly from the codebase of \code{\link{ape}}.}
\usage{
  tree.to.string(phy, digits = 10)
}
\arguments{
  \item{phy,}{input phylo object}

  \item{digits,}{the number of digits to use when
  formatting branch lengths}
}
\value{
  string, the phylo object in a string representation
}
\description{
  Converts a phylo object into a Newick-format string
  representation. Taken mostly from the codebase of
  \code{\link{ape}}.
}

