\name{tree.total.branch.length}
\alias{tree.total.branch.length}
\title{Returns the total branch length of the tree}
\usage{
  tree.total.branch.length(phylo)
}
\arguments{
  \item{phylo}{input phylo object}
}
\value{
  numeric, the sum of all branch lengths in the tree.
}
\description{
  Returns the total branch length of the tree
}

