\docType{data}
\name{GeomJoinedSegment}
\alias{GeomJoinedSegment}
\title{Create a geom_joinedsegment which joins together segments with 'round' corners. Looks MUCH nicer.
See https://groups.google.com/forum/?fromgroups#!topic/ggplot2/movv0f_MSuY}
\format{proto object 
 $ draw   :function (., data, scales, coordinates, arrow = NULL, ...)   
 $ objname: chr "geom_joinedsegment" 
 parent: proto object 
 .. parent: proto object 
 .. .. parent: proto object}
\usage{
  GeomJoinedSegment
}
\description{
  Create a geom_joinedsegment which joins together segments
  with 'round' corners. Looks MUCH nicer. See
  https://groups.google.com/forum/?fromgroups#!topic/ggplot2/movv0f_MSuY
}
\keyword{datasets}

