\name{tree.plot}
\alias{tree.plot}
\title{Plots a tree or a list of trees.}
\usage{
  tree.plot(x, plot.ancestors = T, keep.aspect.ratio = F,
    radial.plot = F, do.plot = T, show.y.gridlines = F,
    plot.legend = T, xlim = NA, xlim.expand = c(0.05, 0.3),
    xlab = "Branch Length", labels.in.axis = F,
    labels.in.tree = T, node.plot = FALSE,
    node.scale.by = NA, node.size = 1, node.color.by = NA,
    node.color = "black", label.plot = TRUE,
    label.scale.by = NA, label.size = 4,
    label.color.by = NA, label.color = "black", y.lim = NA,
    ...)
}
\description{
  Plots a tree or a list of trees.
}

